CRUD update 2

+++ spring-crud/src/main/java/com/spring/crud/spring/crud/service/ProductService.java
@@ -6,13 +6,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
 
     @Autowired
     private ProductRepository productRepository;
-
 
     //service to Post a single product to database
     public Product saveProduct(Product product){
@@ -38,21 +38,27 @@ public class ProductService {
     }
 
     //service to delete a product by product id.
-    public String deleteProduct(int id){
-        productRepository.deleteById(id);
-        return "product deleted!" + id;
+    public String deleteProduct(int id) throws Exception {
+        Optional<Product> p = productRepository.findById(id);
+        System.out.println("p.isEmpty() = " + p.isEmpty());
+        if (!p.isEmpty()) {
+            productRepository.deleteById(id);
+            return "product deleted. id=" + id;
+        }
+        return "product not found with id=" + id;
     }
 
     //service to update a product on database
     public Product updateProduct(Product product){
         //first catch the product by that products id
       Product existingProduct = productRepository.findById(product.getId()).orElse(null);
-//        as there is no Update() method in JPA, we need to get the new product name,
-//         and then set that name to the previously existed product.
+//        as there is no Update() method in JPA, we need to get the new product ,
+//         and then set that to the previously existed product.
         existingProduct.setProductName(product.getProductName());
         existingProduct.setProductQuantity(product.getProductQuantity());
         existingProduct.setProductPrice(product.getProductPrice());
-        //the return the new updated product
+        //then return the new updated product
         return productRepository.save(existingProduct);
     }
+
 }
